fix: working demo after rebase
