+++ src/rai_core/rai/communication/ros2/connectors.py
@@ -15,48 +15,27 @@
 import threading
 import time
 import uuid
-from collections import OrderedDict
-from typing import Any, Callable, Dict, List, Literal, Optional, Tuple, Union, cast
+from typing import Any, Callable, List, Literal, Optional, Tuple, Union
 
-import numpy as np
-import rai_interfaces.msg
 import rclpy
 import rclpy.executors
 import rclpy.node
 import rclpy.time
-import rosidl_runtime_py.convert
-from cv_bridge import CvBridge
-from PIL import Image
-from pydub import AudioSegment
-from rai_interfaces.msg import HRIMessage as ROS2HRIMessage_
-from rai_interfaces.msg._audio_message import AudioMessage as ROS2HRIMessage__Audio
 from rclpy.duration import Duration
 from rclpy.executors import MultiThreadedExecutor
 from rclpy.node import Node
 from rclpy.qos import QoSProfile
-from sensor_msgs.msg import Image as ROS2Image
 from tf2_ros import Buffer, LookupException, TransformListener, TransformStamped
 
-from rai.communication import (
-    ARIConnector,
-    ARIMessage,
-    HRIConnector,
-    HRIMessage,
-    HRIPayload,
-)
+from rai.communication import ARIConnector, HRIConnector
 from rai.communication.ros2.api import (
     ConfigurableROS2TopicAPI,
     ROS2ActionAPI,
-    ROS2ARIMessage,
-    ROS2HRIMessage,
     ROS2ServiceAPI,
     ROS2TopicAPI,
+    TopicConfig,
 )
-
-
-class ROS2ARIMessage(ARIMessage):
-    def __init__(self, payload: Any, metadata: Optional[Dict[str, Any]] = None):
-        super().__init__(payload, metadata)
+from rai.communication.ros2.messages import ROS2ARIMessage, ROS2HRIMessage
 
 
 class ROS2ARIConnector(ARIConnector[ROS2ARIMessage]):
@@ -210,62 +189,6 @@ class ROS2ARIConnector(ARIConnector[ROS2ARIMessage]):
         self._topic_api.shutdown()
         self._executor.shutdown()
         self._thread.join()
-
-
-class ROS2HRIMessage(HRIMessage):
-    def __init__(self, payload: HRIPayload, message_author: Literal["ai", "human"]):
-        super().__init__(payload, message_author)
-
-    @classmethod
-    def from_ros2(
-        cls, msg: rai_interfaces.msg.HRIMessage, message_author: Literal["ai", "human"]
-    ):
-        cv_bridge = CvBridge()
-        images = [
-            cv_bridge.imgmsg_to_cv2(img_msg, "rgb8")
-            for img_msg in cast(List[ROS2Image], msg.images)
-        ]
-        pil_images = [Image.fromarray(img) for img in images]
-        audio_segments = [
-            AudioSegment(
-                data=audio_msg.audio,
-                frame_rate=audio_msg.sample_rate,
-                sample_width=2,  # bytes, int16
-                channels=audio_msg.channels,
-            )
-            for audio_msg in msg.audios
-        ]
-        return ROS2HRIMessage(
-            payload=HRIPayload(text=msg.text, images=pil_images, audios=audio_segments),
-            message_author=message_author,
-        )
-
-    def to_ros2_dict(self) -> OrderedDict[str, Any]:
-        cv_bridge = CvBridge()
-        assert isinstance(self.payload, HRIPayload)
-        img_msgs = [
-            cv_bridge.cv2_to_imgmsg(np.array(img), "rgb8")
-            for img in self.payload.images
-        ]
-        audio_msgs = [
-            ROS2HRIMessage__Audio(
-                audio=audio.raw_data,
-                sample_rate=audio.frame_rate,
-                channels=audio.channels,
-            )
-            for audio in self.payload.audios
-        ]
-
-        return cast(
-            OrderedDict[str, Any],
-            rosidl_runtime_py.convert.message_to_ordereddict(
-                ROS2HRIMessage_(
-                    text=self.payload.text,
-                    images=img_msgs,
-                    audios=audio_msgs,
-                )
-            ),
-        )
 
 
 class ROS2HRIConnector(HRIConnector[ROS2HRIMessage]):
